Renamed community package.
This commit renames the community package. Community is a pretty weird
name, especially if we ever want to use this internally.
