+++ apps/apps.go
@@ -3,14 +3,14 @@
 package apps
 
 import (
-	"tidbyt.dev/community-apps/apps/community"
 	"tidbyt.dev/community-apps/apps/fuzzyclock"
+	"tidbyt.dev/community-apps/apps/manifest"
 )
 
-// GetApps returns a list of all apps in the this repository. Add your applet
+// GetManifests returns a list of all apps in the this repository. Add your applet
 // below to include it in the Tidbyt Mobile app for all Tidbyt users.
-func GetApps() []community.App {
-	return []community.App{
+func GetManifests() []manifest.Manifest {
+	return []manifest.Manifest{
 		fuzzyclock.New(),
 	}
 }
