+++ apps/manifest/testdata/source.star
